Add files via upload
This is the first meeting report. This report occurred when we haven't had anything set up in terms of team communication and project management. The report should contain all of the relevant information for the team and the project, such as team members info, team leader's name, scrum master's name, which tools we have decided to choose for project management and repository, date and time for weekly meeting.